Add: ルーム接続/未接続の切り替えを追加
