+++ src/components/organisms/ChatJoin/component.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const ChatJoin: React.VFC = () => {
+  return <div />;
+};
