+++ src/components/organisms/ChatJoin/index.ts
@@ -1,0 +1,1 @@
+export { ChatJoin } from "./component";
