+++ src/components/organisms/StreamingVideoChat/component.tsx
@@ -1,0 +1,21 @@
+import "aframe";
+import React from "react";
+import { Scene } from "aframe-react";
+
+// Components
+import { ASky } from "../../atoms/ASky";
+import { ThreeScreenVideo } from "../../molecules/ThreeScreenVideo";
+
+// Hooks
+import { useStreamingVideoChatHooks } from "./hooks";
+
+export const StreamingVideoChat = () => {
+  const { threeVideoSources } = useStreamingVideoChatHooks();
+
+  return (
+    <Scene>
+      <ASky material="color: #cccccc" />
+      <ThreeScreenVideo sources={threeVideoSources} />
+    </Scene>
+  );
+};
