+++ src/components/organisms/StreamingVideoChat/index.ts
@@ -1,0 +1,2 @@
+export { StreamingVideoChat } from "./component";
+export { useStreamingVideoChatHooks } from "./hooks";
