+++ src/components/pages/App/hooks.tsx
@@ -1,62 +1,20 @@
-import React, { useEffect, useMemo, useState } from "react";
+import { useEffect, useMemo } from "react";
 import Peer from "skyway-js";
 
-import { AVideoProps } from "../../atoms/AVideo";
-import { AVideoStreamProps } from "../../atoms/AVideoStream";
-export type StreamType = null | MediaStream;
-
 export const useAppHooks = () => {
-  const [stream, setStream] = useState<StreamType>(null);
-
   const peer = useMemo(() => {
     return process.env.REACT_APP_SKYWAY_API_KEY
       ? new Peer({ key: process.env.REACT_APP_SKYWAY_API_KEY as string })
       : null;
   }, [process.env.REACT_APP_SKYWAY_API_KEY]);
 
-  const threeVideoSources = useMemo<(AVideoProps | AVideoStreamProps)[]>(
-    () => [
-      {
-        src: stream,
-        position: "-4 1.5 -5",
-        rotation: "0 45 0",
-        width: "4",
-        height: "2.25",
-      },
-      {
-        src: stream,
-        position: "0 1.5 -5",
-        rotation: "0 0 0",
-        width: "4",
-        height: "2.25",
-      },
-      {
-        src: stream,
-        position: "4 1.5 -5",
-        rotation: "0 -45 0",
-        width: "4",
-        height: "2.25",
-      },
-    ],
-    [stream]
-  );
+  const isJoinRoom = useMemo(() => {
+    return peer ? !!Object.keys(peer.rooms).length : false;
+  }, [peer]);
 
-  // カメラを使用する
   useEffect(() => {
-    (async () => {
-      try {
-        const stream = await navigator.mediaDevices.getUserMedia({
-          video: true,
-          audio: true,
-        });
+    console.log(isJoinRoom);
+  }, [isJoinRoom]);
 
-        setStream(stream);
-      } catch (error) {
-        setStream(null);
-        console.error("mediaDevice.getUserMedia() error", error);
-      }
-    })();
-  }, []);
-
-  return { peer, threeVideoSources, stream };
+  return { peer, isJoinRoom };
 };
