+++ src/components/pages/App/index.ts
@@ -1,2 +1,1 @@
 export { App } from "./component";
-export { useAppHooks } from "./hooks";
